requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=2.0.1
pytest>=7.0.0 